Set the DEBIAN_FRONTEND environment var for shell provisioners
Setting this to "noninteractive" is required to ensure packages are configured fully when installed by the provisioning scripts. The parameter set in the boot command only affects packages installed by the Debian installer. Since we are using the server ISO for the desktop build, most of the desktop packages get installed by the desktop.sh provisioning script and need this variable set. Without it, some packages were failing with the following error:

    debconf: unable to initialize frontend: Dialog
    debconf: (Dialog frontend will not work on a dumb terminal, an emacs shell buffer, or without a controlling terminal.)
    debconf: falling back to frontend: Readline